Create `features/entity-management` directory
