+++ web-common/src/features/entity-management/entity-mappers.ts
@@ -1,4 +1,4 @@
-import { EntityType } from "@rilldata/web-common/lib/entity";
+import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
 
 export function getFilePathFromNameAndType(
   name: string,
