+++ web-common/src/features/entity-management/file-artifacts-store.ts
@@ -1,7 +1,7 @@
 import type { V1ReconcileError } from "@rilldata/web-common/runtime-client";
 import { Readable, writable } from "svelte/store";
 import { parseDocument } from "yaml";
-import type { MetricsConfig } from "./metrics-internal-store";
+import type { MetricsConfig } from "../../../../web-local/src/lib/application-state-stores/metrics-internal-store";
 
 export type FileArtifactsData = {
   // name used for the file. will not always be dependent on the file
