+++ web-common/src/features/models/workspace/ModelWorkspace.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { EntityType } from "@rilldata/web-common/lib/entity";
+  import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
   import { appStore } from "@rilldata/web-local/lib/application-state-stores/app-store";
   import WorkspaceContainer from "@rilldata/web-local/lib/components/workspace/core/WorkspaceContainer.svelte";
   import ModelInspector from "./inspector/ModelInspector.svelte";
