+++ web-common/src/features/models/workspace/ModelWorkspaceHeader.svelte
@@ -4,27 +4,25 @@
   import HideBottomPane from "@rilldata/web-common/components/icons/HideBottomPane.svelte";
   import { notifications } from "@rilldata/web-common/components/notifications";
   import SlidingWords from "@rilldata/web-common/components/tooltip/SlidingWords.svelte";
-  import { EntityType } from "@rilldata/web-common/lib/entity";
+  import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
+  import { fileArtifactsStore } from "@rilldata/web-common/features/entity-management/file-artifacts-store";
   import { useRuntimeServiceRenameFileAndReconcile } from "@rilldata/web-common/runtime-client";
   import {
-    runtimeStore,
     appQueryStatusStore,
+    runtimeStore,
   } from "@rilldata/web-local/lib/application-state-stores/application-store";
-  import { fileArtifactsStore } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store";
   import PanelCTA from "@rilldata/web-local/lib/components/panel/PanelCTA.svelte";
   import { WorkspaceHeader } from "@rilldata/web-local/lib/components/workspace";
-  import {
-    isDuplicateName,
-    renameFileArtifact,
-    useAllNames,
-  } from "@rilldata/web-local/lib/svelte-query/actions";
   import { useGetDashboardsForModel } from "@rilldata/web-local/lib/svelte-query/dashboards";
+  import { useQueryClient } from "@sveltestack/svelte-query";
+  import { getContext } from "svelte";
+  import { renameFileArtifact } from "../../entity-management/actions";
   import {
     getFilePathFromNameAndType,
     getRouteFromName,
-  } from "@rilldata/web-local/lib/util/entity-mappers";
-  import { useQueryClient } from "@sveltestack/svelte-query";
-  import { getContext } from "svelte";
+  } from "../../entity-management/entity-mappers";
+  import { useAllNames } from "../../entity-management/selectors";
+  import { isDuplicateName } from "../../entity-management/utils";
   import ModelWorkspaceCTAs from "./ModelWorkspaceCTAs.svelte";
 
   export let modelName: string;
