+++ web-common/src/features/sources/add-source/createSource.ts
@@ -1,13 +1,13 @@
 import { goto } from "$app/navigation";
 import { notifications } from "@rilldata/web-common/components/notifications";
-import { EntityType } from "@rilldata/web-common/lib/entity";
+import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
+import { getFilePathFromNameAndType } from "@rilldata/web-common/features/entity-management/entity-mappers";
+import { fileArtifactsStore } from "@rilldata/web-common/features/entity-management/file-artifacts-store";
 import type {
   V1PutFileAndReconcileResponse,
   V1ReconcileError,
 } from "@rilldata/web-common/runtime-client";
-import { fileArtifactsStore } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store";
 import { invalidateAfterReconcile } from "@rilldata/web-local/lib/svelte-query/invalidation";
-import { getFilePathFromNameAndType } from "@rilldata/web-local/lib/util/entity-mappers";
 import type { QueryClient, UseMutationResult } from "@sveltestack/svelte-query";
 
 export async function createSource(
