+++ web-common/src/features/sources/navigation/TableAssets.svelte
@@ -1,5 +1,7 @@
 <script lang="ts">
   import { page } from "$app/stores";
+  import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
+  import RenameAssetModal from "@rilldata/web-common/features/entity-management/RenameAssetModal.svelte";
   import {
     useEmbeddedSources,
     useSourceNames,
@@ -13,11 +15,9 @@
   import ColumnProfile from "@rilldata/web-local/lib/components/column-profile/ColumnProfile.svelte";
   import NavigationEntry from "@rilldata/web-local/lib/components/navigation/NavigationEntry.svelte";
   import NavigationHeader from "@rilldata/web-local/lib/components/navigation/NavigationHeader.svelte";
-  import RenameAssetModal from "@rilldata/web-local/lib/components/navigation/RenameAssetModal.svelte";
   import { useQueryClient } from "@sveltestack/svelte-query";
   import { flip } from "svelte/animate";
   import { slide } from "svelte/transition";
-  import { EntityType } from "../../../lib/entity";
   import { useModelNames } from "../../models/selectors";
   import AddSourceModal from "../add-source/AddSourceModal.svelte";
   import { createModelFromSource } from "../createModel";
