+++ web-common/src/features/sources/workspace/SourceWorkspace.svelte
@@ -1,9 +1,11 @@
 <script lang="ts">
   import { Button } from "@rilldata/web-common/components/button";
   import { Callout } from "@rilldata/web-common/components/callout";
+  import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
+  import { getFilePathFromNameAndType } from "@rilldata/web-common/features/entity-management/entity-mappers";
+  import { fileArtifactsStore } from "@rilldata/web-common/features/entity-management/file-artifacts-store";
   import { humanReadableErrorMessage } from "@rilldata/web-common/features/sources/add-source/errors.js";
   import { refreshSource } from "@rilldata/web-common/features/sources/refreshSource";
-  import { EntityType } from "@rilldata/web-common/lib/entity";
   import {
     getRuntimeServiceGetCatalogEntryQueryKey,
     useRuntimeServiceGetCatalogEntry,
@@ -14,11 +16,9 @@
   } from "@rilldata/web-common/runtime-client";
   import { appStore } from "@rilldata/web-local/lib/application-state-stores/app-store";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
-  import { fileArtifactsStore } from "@rilldata/web-local/lib/application-state-stores/file-artifacts-store";
   import { overlay } from "@rilldata/web-local/lib/application-state-stores/overlay-store";
   import { ConnectedPreviewTable } from "@rilldata/web-local/lib/components/preview-table/index.js";
   import { WorkspaceContainer } from "@rilldata/web-local/lib/components/workspace/index.js";
-  import { getFilePathFromNameAndType } from "@rilldata/web-local/lib/util/entity-mappers";
   import { useQueryClient } from "@sveltestack/svelte-query";
   import { parseDocument } from "yaml";
   import SourceInspector from "./SourceInspector.svelte";
