+++ web-local/src/lib/components/column-profile/ColumnProfileIcon.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import { DataTypeIcon } from "@rilldata/web-common/components/data-types";
+  import { EntityStatus } from "@rilldata/web-common/features/entity-management/entity";
   import Spinner from "@rilldata/web-common/features/temp/Spinner.svelte";
-  import { EntityStatus } from "@rilldata/web-common/lib/entity";
   import { fade } from "svelte/transition";
   import { LIST_SLIDE_DURATION } from "../../application-config";
