+++ web-local/src/lib/components/workspace/metrics-def/MetricsDefWorkspaceHeader.svelte
@@ -2,18 +2,16 @@
   import { goto } from "$app/navigation";
   import MetricsIcon from "@rilldata/web-common/components/icons/Metrics.svelte";
   import { notifications } from "@rilldata/web-common/components/notifications";
-  import { EntityType } from "@rilldata/web-common/lib/entity";
+  import { renameFileArtifact } from "@rilldata/web-common/features/entity-management/actions";
+  import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
+  import { useAllNames } from "@rilldata/web-common/features/entity-management/selectors";
+  import { isDuplicateName } from "@rilldata/web-common/features/entity-management/utils";
   import { useRuntimeServiceRenameFileAndReconcile } from "@rilldata/web-common/runtime-client";
   import {
     appQueryStatusStore,
     runtimeStore,
   } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import { useQueryClient } from "@sveltestack/svelte-query";
-  import {
-    isDuplicateName,
-    renameFileArtifact,
-    useAllNames,
-  } from "../../../svelte-query/actions";
   import MetricsDefinitionExploreMetricsButton from "../../metrics-definition/MetricsDefinitionExploreMetricsButton.svelte";
   import WorkspaceHeader from "../core/WorkspaceHeader.svelte";
