+++ web-local/src/routes/(application)/dashboard/[name]/+page.ts
@@ -1,10 +1,10 @@
-import { EntityType } from "@rilldata/web-common/lib/entity";
+import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
+import { getFilePathFromNameAndType } from "@rilldata/web-common/features/entity-management/entity-mappers";
 import {
   runtimeServiceGetCatalogEntry,
   runtimeServiceGetFile,
 } from "@rilldata/web-common/runtime-client";
 import { runtimeServiceGetConfig } from "@rilldata/web-common/runtime-client/manual-clients";
-import { getFilePathFromNameAndType } from "@rilldata/web-local/lib/util/entity-mappers";
 import { error, redirect } from "@sveltejs/kit";
 import { CATALOG_ENTRY_NOT_FOUND } from "../../../../lib/errors/messages";
