+++ web-local/src/routes/(application)/dashboard/[name]/edit/+page.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-  import { EntityType } from "@rilldata/web-common/lib/entity";
+  import { EntityType } from "@rilldata/web-common/features/entity-management/entity";
+  import { getFilePathFromNameAndType } from "@rilldata/web-common/features/entity-management/entity-mappers";
   import { useRuntimeServiceGetFile } from "@rilldata/web-common/runtime-client";
   import { runtimeStore } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import { MetricsDefinitionWorkspace } from "@rilldata/web-local/lib/components/workspace";
-  import { getFilePathFromNameAndType } from "@rilldata/web-local/lib/util/entity-mappers";
 
   export let data;
